pyproject.toml
[tool.poetry]
name = "llama-search"
version = "0.1.0"
description = ""
authors = ["Niels <niels.holvoet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
llama-index = "^0.10.20"
llama-index-readers-notion = "^0.1.6"
llama-index-core = "^0.10.21.post1"
llama-index-llms-ollama = "^0.1.2"
llama-index-readers-file = "^0.1.11"
qdrant-client = "^1.8.0"
llama-index-vector-stores-qdrant = "^0.1.4"
llama-index-embeddings-fastembed = "^0.1.4"
sentence-transformers = "^2.5.1"
tomlkit = "^0.12.4"
rich = "^13.7.1"
llama-index-embeddings-huggingface = "^0.1.4"
langdetect = "^1.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
